Sending Email whenever a new user is created and fixed mocha issue
