+++ src/uiTests/MochaTests/.eslintrc.js
@@ -3,7 +3,15 @@ module.exports = {
         "plugin:vue/essential",
         "eslint:recommended"
     ],
-    env : {
-        mocha : true
-    }
+    overrides: [
+        {
+          files: [
+            "**/__tests__/*.{j,t}s?(x)",
+            "**/tests/unit/**/*.spec.{j,t}s?(x)"
+          ],
+          env: {
+            mocha: true
+          }
+        }
+    ]
 }
